// SysWas.Web/ClientApp/src/app/model/cep/endereco.ts
export class Endereco {

  id: number;
  cep: string;
  logradouro: string;
  complemento: string;
  bairro: string;
  localidade: string;
  uf: string;

}





// public long Cep { get; set; }
// public string Logradouro { get; set; }
// public string Bairro { get; set; }

// public int ThisCidadeId { get; set; }
// public virtual Cidade ThisCidade { get; set; }